Remove CUDAerror.h/cpp
